# server/http/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Technical Test KStyle Hub Backend API docs",
        "title": "Technical Test KStyle Hub API Backend",
        "contact": {
            "name": "Fawzi Linggo",
            "email": "fawzilinggo@google.com"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/member": {
            "get": {
                "description": "Get All Member",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Get All Member",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/presenters.GetMemberResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Edit Member By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Edit Member By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.EditMemberByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Member",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Create Member",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/presenters.CreateMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Member By ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Delete Member By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenters.BasicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "presenters.BasicResponse": {
            "description": "General basic response",
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/presenters.CreateEditResponse"
                },
                "error": {
                    "description": "Error will be empty if SUCCESS",
                    "type": "string",
                    "example": "ERROR or NULL"
                },
                "message": {
                    "type": "string",
                    "example": "FAILED or SUCCESS and etc"
                },
                "status": {
                    "description": "Status turn true if process SUCCESS",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "presenters.CreateEditResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "presenters.CreateMemberRequest": {
            "type": "object",
            "required": [
                "gender",
                "skin_color",
                "skin_type",
                "username"
            ],
            "properties": {
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "skin_color": {
                    "type": "string",
                    "example": "white"
                },
                "skin_type": {
                    "type": "string",
                    "example": "oily"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "presenters.EditMemberByIDRequest": {
            "type": "object",
            "required": [
                "gender",
                "id",
                "skin_color",
                "skin_type",
                "username"
            ],
            "properties": {
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "id": {
                    "type": "string"
                },
                "skin_color": {
                    "type": "string",
                    "example": "white"
                },
                "skin_type": {
                    "type": "string",
                    "example": "oily"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "presenters.GetMemberResponse": {
            "type": "object",
            "properties": {
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "id": {
                    "type": "string",
                    "example": "c836c745-e1db-4777-9699-e40fffaa302d"
                },
                "skin_color": {
                    "type": "string",
                    "example": "white"
                },
                "skin_type": {
                    "type": "string",
                    "example": "oily"
                },
                "username": {
                    "type": "string",
                    "example": "fawzi"
                }
            }
        },
...
    }
}